requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.0.0
pydantic-settings>=2.0.0
elasticsearch>=7.17.0
redis>=5.0.0
prometheus-client>=0.12.0
python-dotenv>=0.19.0
click>=8.0.0
langchain>=0.0.200
vllm>=0.0.1
docling @ git+https://github.com/DS4SD/docling.git
python-magic>=0.4.27
pytest>=7.0.0
pytest-asyncio>=0.16.0
httpx>=0.23.0
python-json-logger>=2.0.0
tenacity>=8.0.0
aiohttp>=3.8.0